refactor(node): move log rotater into own module
