+++ sn_node/src/bin/sn_node/log/mod.rs
@@ -1,0 +1,3 @@
+mod appender;
+
+pub use appender::FileRotateAppender;
